Move `reaction_serializers` into `serializers`
