+++ src/researchhub_comment/serializers/rh_comment_serializer.py
@@ -2,7 +2,7 @@ from django.db.models import Sum
 from rest_framework.serializers import SerializerMethodField
 
 from discussion.models import Vote
-from discussion.reaction_serializers import (
+from discussion.serializers import (
     GenericReactionSerializer,
     GenericReactionSerializerMixin,
     VoteSerializer,
